requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
PyJWT
pydantic